Allow search from URL hash `#q=`, and update browser history upon search.
This lets you link directly to a URL like `example.com/#q=beispiel` and
the search will occur automatically.

If you update the search value, this will be reflected in the URL, and
an entry pushed onto the browser history stack, enabling you to navigate
back and forth between searches.

To improve browser navigation and history, the search term is also now
appended to the page title.